requirements.txt
fastapi
uvicorn
python-docx
beautifulsoup4
requests
python-dotenv
google-generativeai